% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorophyll_cycling.R
\name{chlorophyll_cycling_scaled}
\alias{chlorophyll_cycling_scaled}
\title{scaled chlorophyll cycling model}
\usage{
chlorophyll_cycling_scaled(par, data)
}
\arguments{
\item{par}{three parameters specifying the cycling model}

\item{data}{data consisting of vegetation greenness (G), mean
daytime temperature (T) and daylenght (D) as a data frame}
}
\value{
Vegetation greenness values
}
\description{
Models the senescence of vegetation as mass balance of
produced and lost chlorophyll in response to temperature
and daylength, this in comparison to a threshold based
approach which accumulated (chilling) degree days.
}
\details{
Deviates from the original by scaling synthesis with
the state of the previous Gcc value (production capacity is dependent)
on available tissue/capacity in t-1
}
