% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_list.R
\name{df_to_list}
\alias{df_to_list}
\title{Converts a long data frame to a listed model data structure}
\usage{
df_to_list(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\value{
wide format matrix per environmental variable split with columns
named after site and year, and rows as day-of-year
}
\description{
Conversion of a long format data frame of environmental driver data
and PhenoCam reference data to a wide format where every list item
contains a large matrix of each value, with columns the site years and
rows day-of-year.
}
\details{
This allows for vectorization of the certain operations to speed up
model parameterization.
}
