% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_ts.R
\name{normalize_ts}
\alias{normalize_ts}
\title{Normalize PhenoCam time series}
\usage{
normalize_ts(df, prob = c(0.1, 0.9))
}
\arguments{
\item{df}{a PhenoCam Gcc time series}

\item{prob}{probability values (0-1) as vector to denote min
and max values in normalizing}
}
\value{
A normalized PhenoCam time series.
}
\description{
Normalize PhenoCam data between 0-1 to to standardize
further processing using quantile probability values
implicitly addressing outliers
}
